fix code imports and tests for host.ts
